requirements.txt
python-dotenv
streamlit
groq
langchain-community
langchain
pydantic
